backend/src/test/java/com/capstone/ems/domain/dto/EmployeeDtoTests.java
package com.capstone.ems.domain.dto;

import static org.junit.jupiter.api.Assertions.assertEquals;

import java.util.Arrays;
import java.util.List;

import org.junit.jupiter.api.Test;

import com.capstone.ems.enums.UserType;

public class EmployeeDtoTests {

    @Test
    public void testGettersAndSetters() {
        Long empId = 1L;
        String username = "johndoe";
        String email = "john.doe@example.com";
        String name = "John Doe";
        List<String> skills = Arrays.asList("Java", "Python");
        Long managerId = 2L;
        List<Long> managedProjectIds = Arrays.asList(3L, 4L);
        Long projectId = 5L;
        UserType userType = UserType.EMPLOYEE;
        Long userId = 6L;

        EmployeeDto employeeDto = new EmployeeDto();

        employeeDto.setEmpId(empId);
        employeeDto.setUsername(username);
        employeeDto.setEmail(email);
        employeeDto.setName(name);
        employeeDto.setSkills(skills);
        employeeDto.setManagerId(managerId);
        employeeDto.setManagedProjectIds(managedProjectIds);
        employeeDto.setProjectId(projectId);
        employeeDto.setUserType(userType);
        employeeDto.setUserId(userId);

        assertEquals(empId, employeeDto.getEmpId());
        assertEquals(username, employeeDto.getUsername());
        assertEquals(email, employeeDto.getEmail());
        assertEquals(name, employeeDto.getName());
        assertEquals(skills, employeeDto.getSkills());
        assertEquals(managerId, employeeDto.getManagerId());
        assertEquals(managedProjectIds, employeeDto.getManagedProjectIds());
        assertEquals(projectId, employeeDto.getProjectId());
        assertEquals(userType, employeeDto.getUserType());
        assertEquals(userId, employeeDto.getUserId());
    }

    @Test
    public void testBuilder() {
        Long empId = 1L;
        String username = "johndoe";
        String email = "john.doe@example.com";
        String name = "John Doe";
        List<String> skills = Arrays.asList("Java", "Python");
        Long managerId = 2L;
        List<Long> managedProjectIds = Arrays.asList(3L, 4L);
        Long projectId = 5L;
        UserType userType = UserType.EMPLOYEE;
        Long userId = 6L;

        EmployeeDto employeeDto = new EmployeeDto.Builder()
                .setEmpId(empId)
                .setUsername(username)
                .setEmail(email)
                .setName(name)
                .setSkills(skills)
                .setManagerId(managerId)
                .setManagedProjectIds(managedProjectIds)
                .setProjectId(projectId)
                .setUserType(userType)
                .setUserId(userId)
                .build();

        assertEquals(empId, employeeDto.getEmpId());
        assertEquals(username, employeeDto.getUsername());
        assertEquals(email, employeeDto.getEmail());
        assertEquals(name, employeeDto.getName());
        assertEquals(skills, employeeDto.getSkills());
        assertEquals(managerId, employeeDto.getManagerId());
        assertEquals(managedProjectIds, employeeDto.getManagedProjectIds());
        assertEquals(projectId, employeeDto.getProjectId());
        assertEquals(userType, employeeDto.getUserType());
        assertEquals(userId, employeeDto.getUserId());
    }
}